server/requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==2.2.2
dj-database-url==0.4.2
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-webpack-loader==0.5.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
flake8==3.5.0
google-api-python-client==1.6.2
graphene==2.1.2
graphene-django==2.1rc1
graphql-core==2.1rc3
graphql-relay==0.4.5
gunicorn==19.7.1
httplib2==0.11.3
idna==2.6
iso8601==0.1.12
mccabe==0.6.1
oauth2client==4.1.2
pluggy==0.6.0
promise==2.1
psycopg2==2.7.3.1
py==1.5.4
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.5.3
python-env==1.0.0
pytz==2018.4
requests==2.18.4
rsa==3.4.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
slackclient==1.0.9
tox==3.0.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.0.0
websocket-client==0.44.0
whitenoise==3.3.0
